+++ tests/ethercat/test_pdo.py
@@ -198,26 +198,40 @@ def test_servo_add_maps(connect_to_slave, create_pdo_map):
     servo.reset_tpdo_mapping()
     servo.reset_rpdo_mapping()
 
-    assert servo.read(EthercatServo.TPDO_ASSIGN_REGISTER_SUB_IDX_0) == 0
-    assert servo.read(EthercatServo.RPDO_ASSIGN_REGISTER_SUB_IDX_0) == 0
+    assert servo.read(EthercatServo.TPDO_ASSIGN_REGISTER_SUB_IDX_0, subnode=0) == 0
+    assert servo.read(EthercatServo.RPDO_ASSIGN_REGISTER_SUB_IDX_0, subnode=0) == 0
 
     servo.set_pdo_map_to_slave([rpdo_map], [tpdo_map])
     servo.map_pdos(1)
 
-    assert servo.read(EthercatServo.TPDO_ASSIGN_REGISTER_SUB_IDX_0) == 1
+    assert servo.read(EthercatServo.TPDO_ASSIGN_REGISTER_SUB_IDX_0, subnode=0) == 1
     assert len(servo._tpdo_maps) == 1
-    assert tpdo_map.map_register_index == EthercatServo.TPDO_MAP_REGISTER_SUB_IDX_0[0].idx
+    assert (
+        tpdo_map.map_register_index
+        == servo.dictionary.registers(0)[servo.TPDO_MAP_REGISTER_SUB_IDX_0[0]].idx
+    )
     assert tpdo_map.map_register_index_bytes == tpdo_map.map_register_index.to_bytes(4, "little")
-    assert servo.read(EthercatServo.TPDO_MAP_REGISTER_SUB_IDX_0[0]) == len(TPDO_REGISTERS)
-    value = servo._read_raw(EthercatServo.TPDO_ASSIGN_REGISTER_SUB_IDX_0, complete_access=True)
+    assert servo.read(EthercatServo.TPDO_MAP_REGISTER_SUB_IDX_0[0], subnode=0) == len(
+        TPDO_REGISTERS
+    )
+    value = servo._read_raw(
+        servo.dictionary.registers(0)[servo.TPDO_ASSIGN_REGISTER_SUB_IDX_0], complete_access=True
+    )
     assert int.to_bytes(0x1A00, 2, "little") == value[2:4]
 
-    assert servo.read(EthercatServo.RPDO_ASSIGN_REGISTER_SUB_IDX_0) == 1
+    assert servo.read(EthercatServo.RPDO_ASSIGN_REGISTER_SUB_IDX_0, subnode=0) == 1
     assert len(servo._rpdo_maps) == 1
-    assert rpdo_map.map_register_index == EthercatServo.RPDO_MAP_REGISTER_SUB_IDX_0[0].idx
+    assert (
+        rpdo_map.map_register_index
+        == servo.dictionary.registers(0)[servo.RPDO_MAP_REGISTER_SUB_IDX_0[0]].idx
+    )
     assert rpdo_map.map_register_index_bytes == rpdo_map.map_register_index.to_bytes(4, "little")
-    assert servo.read(EthercatServo.RPDO_MAP_REGISTER_SUB_IDX_0[0]) == len(RPDO_REGISTERS)
-    value = servo._read_raw(EthercatServo.RPDO_ASSIGN_REGISTER_SUB_IDX_0, complete_access=True)
+    assert servo.read(EthercatServo.RPDO_MAP_REGISTER_SUB_IDX_0[0], subnode=0) == len(
+        RPDO_REGISTERS
+    )
+    value = servo._read_raw(
+        servo.dictionary.registers(0)[servo.RPDO_ASSIGN_REGISTER_SUB_IDX_0], complete_access=True
+    )
     assert int.to_bytes(0x1600, 2, "little") == value[2:4]
 
 
@@ -229,16 +243,16 @@ def test_servo_reset_pdos(connect_to_slave, create_pdo_map):
     servo.set_pdo_map_to_slave([rpdo_map], [tpdo_map])
     servo.map_pdos(1)
 
-    assert servo.read(EthercatServo.TPDO_ASSIGN_REGISTER_SUB_IDX_0) == 1
-    assert servo.read(EthercatServo.RPDO_ASSIGN_REGISTER_SUB_IDX_0) == 1
+    assert servo.read(servo.TPDO_ASSIGN_REGISTER_SUB_IDX_0, subnode=0) == 1
+    assert servo.read(servo.RPDO_ASSIGN_REGISTER_SUB_IDX_0, subnode=0) == 1
     assert len(servo._rpdo_maps) == 1
     assert len(servo._tpdo_maps) == 1
 
     servo.reset_tpdo_mapping()
     servo.reset_rpdo_mapping()
 
-    assert servo.read(EthercatServo.TPDO_ASSIGN_REGISTER_SUB_IDX_0) == 0
-    assert servo.read(EthercatServo.RPDO_ASSIGN_REGISTER_SUB_IDX_0) == 0
+    assert servo.read(servo.TPDO_ASSIGN_REGISTER_SUB_IDX_0, subnode=0) == 0
+    assert servo.read(servo.RPDO_ASSIGN_REGISTER_SUB_IDX_0, subnode=0) == 0
     assert len(servo._rpdo_maps) == 0
     assert len(servo._tpdo_maps) == 0
 
@@ -365,15 +379,15 @@ def test_set_pdo_map_to_slave(connect_to_slave, create_pdo_map):
 
     servo.map_pdos(1)
     # Check the current PDO mapping
-    assert servo.read(EthercatServo.TPDO_ASSIGN_REGISTER_SUB_IDX_0) == 1
-    assert servo.read(EthercatServo.RPDO_ASSIGN_REGISTER_SUB_IDX_0) == 1
+    assert servo.read(servo.TPDO_ASSIGN_REGISTER_SUB_IDX_0, subnode=0) == 1
+    assert servo.read(servo.RPDO_ASSIGN_REGISTER_SUB_IDX_0, subnode=0) == 1
 
     new_rdpo_map = RPDOMap()
     new_tpdo_map = TPDOMap()
     servo.set_pdo_map_to_slave([new_rdpo_map], [new_tpdo_map])
     # Check that the previous mapping was not deleted
-    assert servo.read(EthercatServo.TPDO_ASSIGN_REGISTER_SUB_IDX_0) == 1
-    assert servo.read(EthercatServo.RPDO_ASSIGN_REGISTER_SUB_IDX_0) == 1
+    assert servo.read(servo.TPDO_ASSIGN_REGISTER_SUB_IDX_0, subnode=0) == 1
+    assert servo.read(servo.RPDO_ASSIGN_REGISTER_SUB_IDX_0, subnode=0) == 1
     # Check that the new PDOMaps were added
     assert len(servo._rpdo_maps) == 2
     assert servo._rpdo_maps[1] == new_rdpo_map